% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fData.R
\docType{data}
\name{flightData}
\alias{flightData}
\title{Data Collected From Alaska Airlines}
\format{An object of class \code{data.frame} with 714 rows and 20 columns.}
\usage{
flightData
}
\description{
Data Collected From Alaska Airlines
}
\references{
\url{http://users.stat.umn.edu/~almquist/3811_examples/all_alaska_flights.csv}
}
\author{
Alaska Airlines
}
\keyword{datasets}
