% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwQuizFunctions.R
\name{matMult}
\alias{matMult}
\title{Matrix & Vector Computation}
\usage{
matMult(mat, vec)
}
\arguments{
\item{mat}{A 2-D Matrix}

\item{vec}{A numerical vector}
}
\value{
Scalar
}
\description{
Will compute (x)(matrix)(x^T)
}
\examples{
load(url("http://www.stat.umn.edu/geyer/3701/data/q2p1.rda"))
matMult(a, x)

}
