% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q1p3Data.R
\docType{data}
\name{q1p3}
\alias{q1p3}
\title{Example data - HW 1}
\format{An object of class \code{data.frame} with 5 rows and 5 columns.}
\usage{
q1p3
}
\description{
Example data - HW 1
}
\references{
\url{http://www.stat.umn.edu/geyer/3701/data/q1p3.txt}
}
\author{
Charlie Geyer \email{geyer@umn.edu}
}
\keyword{datasets}
