% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwQuizFunctions.R
\name{stdize}
\alias{stdize}
\title{Standardize A 2-D Matrix}
\usage{
stdize(dat)
}
\arguments{
\item{dat}{A 2-D Matrix containing numerical values}
}
\value{
A 2-D Matrix
}
\description{
Will compute a standardized matrix by column.
}
\examples{
load(url("http://www.stat.umn.edu/geyer/3701/data/q2p1.rda"))
stdize(a)

}
